Fully implement ReadCountLE/WriteCountLE ReadBytes/WriteBytes
I played with implementing ReadCountBE/WriteCountBE but because of misuse potential i opted to not move forwards with the implementation. The main issue is that such an implementation allows passing unmanaged structs and they just aren't possible to flip the bytes properly on a per-field basis.

This also fixes a few minor bugs with running in debug mode (a bad assertion) and optimizes quite a few cases in the implementation of these functions.